Adding support for federated urls in modals (#2561)
* Adding support for federated urls in modals

* Condtional impot of federated function in modal

* Updating test cases for federated.js
